+++ tool/ci_setup.sh
@@ -13,7 +13,7 @@ case "${1:?usage: tool/ci_setup.sh build|e2e|eval|train}" in
   build) apt clang cmake ninja-build pkg-config libgtk-3-dev liblzma-dev ;;
   e2e)
     apt libgtk-3-dev xvfb xdotool imagemagick sqlite3 openbox wmctrl x11-utils \
-      zip unzip bc desktop-file-utils xdg-utils
+      zip unzip bc desktop-file-utils xdg-utils ffmpeg
     pip pillow numpy opencv-python-headless pypdfium2 ;;
   eval) pip opencv-python-headless numpy pillow pypdfium2 onnxruntime ;;
   train)